refactor(orbit): split package by responsibility, one test file per source
Reorganize internal/tools/orbit/ so every .go file has a matching _test.go
file under the same name, matching the project convention used by the
other 176 tool sub-packages.

- Extract query-DSL validators (validateQuery, requireScopedNodes,
  collectQueryNodes, nodeHasScope, toInt64, requireNeighborsShape,
  requireAtLeastTwoNodes, queryType, queryTypeAlternatives) from orbit.go
  into a dedicated validators.go. validators_test.go already covers every
  helper, so it now sits next to its source.
- Move action_specs and RegisterMeta tests out of orbit_test.go into a
  new action_specs_test.go, together with their shared helpers
  (orbitActionSpecRoutes, registeredToolNames, newOrbitMetaSession,
  newOrbitSpecsByTool, registerOrbitMetaForTest, newOrbitMCPSession,
  containsTool).
- Move markdown formatter tests out of orbit_test.go into a new
  markdown_test.go. TestOrbitMarkdownAndDynamicJSON_FallbackBranches
  stays in orbit_test.go because it also covers the convert* helpers
  and decodeRaw from orbit.go.
- Move the misplaced Status/Schema/GraphStatusOptions/ActionSpecs
  tests from validators_test.go to orbit_test.go/action_specs_test.go,
  and drop the now-unused imports from validators_test.go.

orbit.go shrinks from 1010 to 818 lines and every section now has a
test file named after it. Coverage stays at 100%.
